Move remaining control sys tags to Tags/ directory
