+++ src/ControlSystem/Tags/FunctionsOfTimeInitialize.hpp
@@ -10,8 +10,8 @@
 #include <unordered_map>
 
 #include "ControlSystem/ExpirationTimes.hpp"
-#include "ControlSystem/Tags.hpp"
 #include "ControlSystem/Tags/OptionTags.hpp"
+#include "ControlSystem/Tags/SystemTags.hpp"
 #include "DataStructures/DataBox/DataBox.hpp"
 #include "Domain/Creators/OptionTags.hpp"
 #include "Domain/Creators/Tags/FunctionsOfTime.hpp"
