+++ tests/Unit/ControlSystem/ControlErrors/Test_Shape.cpp
@@ -18,9 +18,9 @@
 #include "ControlSystem/ControlErrors/Shape.hpp"
 #include "ControlSystem/Controller.hpp"
 #include "ControlSystem/Systems/Shape.hpp"
-#include "ControlSystem/Tags.hpp"
 #include "ControlSystem/Tags/MeasurementTimescales.hpp"
 #include "ControlSystem/Tags/QueueTags.hpp"
+#include "ControlSystem/Tags/SystemTags.hpp"
 #include "ControlSystem/TimescaleTuner.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/LinkedMessageQueue.hpp"
