+++ tests/Unit/ControlSystem/Systems/Test_Expansion.cpp
@@ -9,8 +9,8 @@
 #include <tuple>
 
 #include "ControlSystem/Systems/Expansion.hpp"
-#include "ControlSystem/Tags.hpp"
 #include "ControlSystem/Tags/MeasurementTimescales.hpp"
+#include "ControlSystem/Tags/SystemTags.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "Domain/CoordinateMaps/CoordinateMap.hpp"
 #include "Domain/CoordinateMaps/CoordinateMap.tpp"
