+++ tests/Unit/ControlSystem/Systems/Test_Shape.cpp
@@ -18,8 +18,8 @@
 #include "ControlSystem/ControlErrors/Shape.hpp"
 #include "ControlSystem/Controller.hpp"
 #include "ControlSystem/Systems/Shape.hpp"
-#include "ControlSystem/Tags.hpp"
 #include "ControlSystem/Tags/MeasurementTimescales.hpp"
+#include "ControlSystem/Tags/SystemTags.hpp"
 #include "ControlSystem/TimescaleTuner.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/LinkedMessageQueue.hpp"
