+++ tests/Unit/Helpers/ControlSystem/SystemHelpers.hpp
@@ -29,10 +29,10 @@
 #include "ControlSystem/Systems/Rotation.hpp"
 #include "ControlSystem/Systems/Shape.hpp"
 #include "ControlSystem/Systems/Translation.hpp"
-#include "ControlSystem/Tags.hpp"
 #include "ControlSystem/Tags/IsActive.hpp"
 #include "ControlSystem/Tags/MeasurementTimescales.hpp"
 #include "ControlSystem/Tags/QueueTags.hpp"
+#include "ControlSystem/Tags/SystemTags.hpp"
 #include "ControlSystem/TimescaleTuner.hpp"
 #include "ControlSystem/UpdateControlSystem.hpp"
 #include "DataStructures/DataBox/Tag.hpp"
